Fixed offset calculation in iget

+++ lightfs.h
@@ -122,7 +122,7 @@ char *get_block(struct super_block *sb, __u32 num);
 char *blkcpy(struct buffer_head **bh, struct lightfs_superblock *sbi);
 struct buffer_head **get_block_bh(struct super_block *sb, __u32 num);
 
-struct inode *lightfs_iget(struct super_block *sb, size_t inode);
+struct inode *lightfs_iget(struct super_block *sb, unsigned long inode);
 int write_inode(struct inode *inode, __u32 ino);
 
 int init_dir(struct super_block *sb, struct inode *dir, struct inode *parent);
